Removed an unnecessary cover line that caused problems in some cases.
The original path ('File') that was given from cover:is_compiled/1 can
sometimes be a relative path, and if that path was given to cover:compile/1 it
would become a full path in a subsequent call to cover:is_compiled/1 and thus
would not match. The line itself is unnecessary, it is basically an assert
that cover:compile/1 was successful.