Record cover data on passtrhough calls
If a module with cover instrumentation is mocked then make sure to
instrument any passtrhough calls.  This way coverage analysis is still
available for passtrhough calls.

This functionality can be disabled via the `no_passthrough_cover`
option.

Implementation Details
----------------------

This is coded for the specific use case of running eunit tests from
rebar with coverage analysis enabled.  Previously, if just some of the
functions in a src module were mocked then all coverage analyis on
said module, while it was mocked, was lost.

1. check if module is instrumented for coverage

2. compile `<name>_meck_original` (thus known as `OriginalMod`) and
   then cover compile it

3. let meck do it's thing

4. during unload/termination of mocked module make sure to first
   export the cover data collected on `OriginalMod`, after exporting
   modify the data to use the real modules name
   (e.g. `foo_meck_original` -> `foo`), this way the data can be
   imported and counted against `foo`

5. let meck do it's usual cleanup

6. during restore import the `OriginalMod` coverage data

Tricks are played with cover's BEAM code so that private functions can
be called.  This was done to avoid creating temporary files and
copy/pasting code from cover.