Erl19 Breaks Mecking Callbacks/Behaviour Modules
This PR adds a single test with two additional helper modules used by the test.

The test demonstrates mocking both a behavior and callback module in the same
meck:new/? call.  This test passes in erlang 18.x as expected, but when running
from erlang 19, this test fails.

The error indicates that some compiler is not seeing the mocked behavior module
as the behavior module that it is.  I don't have a definitive reason for this,
but I have strong suspicions that the reason is related to:

http://erlang.org/pipermail/erlang-questions/2016-June/089615.html
http://erlang.org/pipermail/erlang-questions/2017-January/091396.html

Basically in erlang 19 the call M:module_info() seems to have stopped returning
the list of callbacks as attributes.  My guess is that Meck is using these in
some way to create the mocked version of the behavior module.  As the email
list posts state the callbacks are still there via beam_lib, so if a fix is not
forthcoming from erlang itself, it may be possible to fix this by using beam_lib,
or perhaps by using merl to examine the AST.