Isolate backup *.coverdata from other beam instances
If multiple beam instances are launched in parallel with the same current
working directory, it is possible for the backup cover data from one
instance to be squashed and deleted by other instances. By adding the OS
pid of beam.smp itself to the filename, we avoid this potential issue.