Increase meck_proc stop timeout to infinity
When stopping (e.g. due to meck:unload/1), meck_procs restore
the original module, which may be a time consuming operation.
If a gen_server:call is used for stopping, there's risk of a
timeout that will be translated to a confusing
error:{not_mocked, ...}.

Using gen_server:stop/1 breaks tests, so using an infinity timeout
instead.