Fix an issue with the code coverage in Erlang.mk
It seems that just purging away the meck generated module
(in `meck:unload`) is not removing it from the cover data.
Exporting the cover data using `cover:export/1` will also
export data about the meck generated module. Erlang.mk
[uses](https://github.com/ninenines/erlang.mk/blob/b8a27ab752389394bfdb8c15d1b69455c313991e/plugins/cover.mk#L52)
the export & import flow to generate the HTML coverage report,
so the report will show the purged module - which is not correct.

This might be due to an issue in the coverage tool and it would be
better to investigate and fix it there, but this fix shouldn't break
anything, so I think it is acceptable.