Add meck:wait_for/6
When testing some non-deterministic code, I needed something that would
wait for a function to be called with [1, 2, 3], but not necessarily all
at once, and not necessarily in that order.

That is: m:f([1]), m:f([2]), m:f([3]) is just as valid as m:f([1, 2]),
m:f([3]), and as valid as m:f([2]), m:f([3, 1]), and so on.

So: invent a way for meck to update what it's waiting for as the history
is built. By accumulating state as the calls happen, we can check that
the function is called with the correct arguments, whatever the order.

This is called a 'condition'.

I also succeeded in generalising 'times_called' to demonstrate its
usefulness.